spotify-producer/requirements.txt
spotipy>=2.25.0
python-dotenv>=1.0.0
kafka-python>=2.0.2
